requirements.txt
python-telegram-bot==20.7
rembg[cpu]==2.0.56
